Only delete Local Storage for websites.
Both CCleaner and Bleachbit deletes the files in `Local Storage` but only the ones that start with `http`.
This is probably because chrome extensions sometimes store their settings in files named like this: `chrome-extension_*.localstorage`
If you delete them you delete all the local settings for those extensions will be gone, I had personally lots of login problems with Lastpass because of this. 

I noticed that you have excluded Lastpass cookies from CCleaner so I figured that these changes would make sense. (But doesn't bleachbit delete those cookies anyway?)

If this isn't a fit for tron then consider it for the standalone temfilecleanup script.