Update prerun_checks_and_tasks.bat
I encountered this error message while trying to execute from c:\tron.  
This location is NOT mentioned in the documentation as an invalid location.

Since I was running this blind on a remote system that will not have access to user network resources while in safe mode, following the instructions would result in a failed execution. 

The attitude was not appreciated.

I would suggest not calling the user a clown and being abrasive when avoidable.
It turns a small issue that's easily fixed into yet another aggravating point of friction.  

This is actually somewhat more important in this case, considering that people don't run tron when they're having a good day. 

Pissing off users when they have an issue just makes them LESS capable of fixing the issue themselves and increases the likelihood of them not understanding the instructions if they DO go re-read them. 

Then you're left with upset users complaining on reddit or opening spurious bug reports for minor things that they could easily fix themselves..

The really bad ones might even submit a pull request. <grin>