Fix disk space calculation under W11
Find either of the matching words "avail" or "free" (as used to be the case), prepend each line with "[#]" and then return the first literal match of `[1]` (which as the first line, is presumed to be the free space output). We then strip spaces and commas, which is needed for the FREE_SPACE_BEFORE calculation.